Refactor: rename ChartGraphX
- Renamed instances of ChartGraphX to CartoGraphX.
