+++ README.md
@@ -1,13 +1,13 @@
-# ChartGraphX VS Code extension
+# CartoGraphX VS Code extension
 
-The ChartGraphX VS Code extension is used to track and share user-metrics inside a team workspace to be viewed on the interactive source-code map displayed on our website. This tool will greatly increase accountability because each developer’s work will be visible in real time.
+The CartoGraphX VS Code extension is used to track and share user-metrics inside a team workspace to be viewed on the interactive source-code map displayed on our website. This tool will greatly increase accountability because each developer’s work will be visible in real time.
 
 ## Running the extension
 
 To run the extension you have clone the repo and open the extension folder in VS Code.    
-``git clone https://github.com/cse112-sp20/ChartGraphX-MVP-Mini-Project-2-.git``   
-``cd ChartGraphX-MVP-Mini-Project-2-/extension``   
-``npm i``  
+``git clone https://github.com/cse112-sp20/CartoGraphX-VS-Extension.git``   
+``cd CartoGraphX-VS-Extension``	
+``npm i``    
 ``code .``    
 
 
@@ -18,11 +18,11 @@ Run the extension command by pressing the CGphX button on your status bar (botto
 
 ## Features
 <img align="right" src="/readme.img/CGphx_proto.png" alt="prototypes"
-	title="GhartGraphX extension" width="50%" />
+	title="CartoGraphX" width="50%" />
 To view the extension commands, press the CGphX button on your status bar (bottom bar). This will prompt a dropdown menu in your IDE providing you with the available commands.  
 
 
-ChartGraphX provides commands to sign up/in/out of the service using email and password. The color of the CGphX button indicates wether the extension is active or not.   
+CartoGraphX provides commands to sign up/in/out of the service using email and password. The color of the CGphX button indicates wether the extension is active or not.   
 
 When signed in, you have the ability to check user credentials, as well as running the displayCurrentWorkingFile() command.
