+++ src/auth.ts
@@ -12,7 +12,7 @@ export const signIn = async (auth: firebase.auth.Auth) => {
     const password = await vscode.window.showInputBox({placeHolder: "password", password: true}); // 'testPassword'
     auth.signInWithEmailAndPassword(email !== undefined ? email : "", password !== undefined ? password : "")
         // tslint:disable-next-line: max-line-length
-        .then( (res) => vscode.window.showInformationMessage("You are now signed to ChartGraphX as: " + auth.currentUser?.email))
+        .then( (res) => vscode.window.showInformationMessage("You are now signed to CartoGraphX as: " + auth.currentUser?.email))
         .catch((error: { code: any; message: any; }) => {
             vscode.window.showWarningMessage(error.message);
         },
@@ -27,7 +27,7 @@ export const signIn = async (auth: firebase.auth.Auth) => {
  */
 export const signOut = (auth: firebase.auth.Auth) => {
     auth.signOut()
-        .then( (res) => vscode.window.showInformationMessage("You are now signed out of ChartGraphX!"))
+        .then( (res) => vscode.window.showInformationMessage("You are now signed out of CartoGraphX!"))
         .catch((error: { code: string; message: string; }) => {
             vscode.window.showWarningMessage(error.message);
     });
