+++ src/cartographx.ts
@@ -31,6 +31,6 @@ export async function createMapFunction() {
 }
 
 export const statusBarItem = vscode.window.createStatusBarItem(undefined, 1000);
-statusBarItem.command = "chartGraphX.toolbarAction";
+statusBarItem.command = "cartoGraphX.toolbarAction";
 statusBarItem.tooltip = "CGphX commands";
 statusBarItem.text = "CGphX$(rocket)"; // to rotate icon ~spin
